new structure for panel visiblity
We no longer use panel.hide() and changing panel height to 0.
Instead we move panel out of display area.

This can help us with having panel corners while panel
isn't visible.

Also hot corner is fixed when panel is disabled because
we don't change the panel height and visiblity.

With this update we can also manage to add an option for
panel visiblity only in overview.