fix override theme and hidden panel compatiblity
Panel may get into the working area when the override theme applies
because the font size can make the panel height larger or smaller
currently API doesn't do it automatically, so we are using
private method form API in Manager (which shouldn't be happening
but this is just a hack. In the future updates API should do
that not Manager).