fix search entry height on no panel
Since we can have margin for search entry parent element, we can
have wrong height in some racy situations. To fix that we always
make the search entry parent height as auto after the first
ease (we cannot ease to -1 since it can cause the jump).

This fix can also solve the entering jump to app grid
when the search entry is disabled.

Related to #19