Fix panel flickering when panel position is at the bottom
`_emitPanelPositionChanged` could cause flickering or recursive (when
fixing looking glass position is getting called).

In this fix, the responsibility of each change get fixed in their own
feature.

For the maximized window going under panel in Wayland, that's a Meta bug
and it should be fixed in Meta not in GNOME Shell extensions:
Mutter/issue#1627

Fixes #195