Fix Looking Glass Resize Error
Replaced the use of `_resize()` for the Looking Glass with the `show` signal.
This approach is less intrusive, as we can modify the `show` signal directly
without requiring corresponding changes to the Looking Glass implementation
for every related change.

This allows us to handle both resizing and positioning within
the `show` signal. It also reduces monkey patching frequency
to twice (for the API's open and close methods).